Try one layer of indirection
